Mute option
added a button on to the main menu to mute out all the sounds. Mute starts
out true with all sounds turned off. Pressing the button will turn the
sounds on and then pressing again to turn them back off.

-added functionality to Hater_MidScreen to move enemy off screen after a
certain number of steps. This is to stay in line with other shmups